casting fix

+++ OverhaulDLL/src/GameData.cpp
@@ -1619,7 +1619,7 @@ void* Game::game_smallObject_malloc(uint64_t heap, size_t size, size_t alignment
 void* Game::thread_malloc(size_t size)
 {
     void* threadObj = (void*)0x141C96ED0;
-    return hkThreadMemory_alloc(threadObj, size);
+    return hkThreadMemory_alloc(threadObj, (uint32_t)size);
 }
 
 void Game::game_free(void* p)
@@ -1645,7 +1645,7 @@ void Game::game_free_alt(void* p, void* heapObjArg)
 void Game::thread_free(void* p, size_t size)
 {
     void* threadObj = (void*)0x141C96ED0;
-    hkThreadMemory_free(threadObj, p, size);
+    hkThreadMemory_free(threadObj, p, (uint32_t)size);
 }
 
 bool Game::Check_DLHeapManager_DLReadWriteLock_IsUnlocked(DWORD timeoutms)
